+++ src/edu/uga/cs/rentaride/presentation/UpdateLateFee.java
@@ -111,7 +111,7 @@ public class UpdateLateFee
         priceS = req.getParameter( "membershipPrice" );
 
         try{
-            price = Integer.valueOf(priceS);
+            price = Double.valueOf(priceS);
         }catch(Exception e) {
 
         }
